# Ifcfile entity look-up + Kernel::get_layers optimizations