Recently we reran conversion for a file ("MEP File") that to our best of recollection was converted in somewhat orderly fashion a long time ago. This time, the XML conversion seemed to take forever, so I started to looking into the biggest time spenders in XML serialization code, which lead mostly to IfcFile::getInverse(). This lead me to optimize entity look-ups and to remove one particular slow part of layer retrieval code that doesn't seem to yield any results ever.

```
Timing includes only the conversion ("-layer code" == unnecessary(?) layer code removed, no other optimizations)
Conference center (http://openifcmodel.cs.auckland.ac.nz/_models/20160124OTC-Conference%20Center.ifc):
before:         1 m 15 s
-layer code:         8 s
after:               7 s

"Massive File" (1.1 GB)
before:         21 m 2 s
-layer code:    19 m 46 s
after:           6 m 58 s

"MEP File" (180 MB)
before:         did not finish (hours?)
-layer code:    14 m 58 s
after:           3 m 25 s
```