@Stinkfist0 2018-10-02
Also ideally, maybe one "entities by ref" map would suffice, and I tried to achieve that, but IfcEntityList made the task rather awkward, so I dediced not to pursue that for now.