@Stinkfist0 2018-10-15
I'll try to to continue with this PR this week.