@Stinkfist0 2018-12-10
I'll finish this this week.