@aothms 2018-12-24
> Clearing the whole cache should not really affect the performance of typical use cases where the file is never modified though.

I agree. 

> one particular slow part of layer retrieval code that doesn't seem to yield any results ever

I ran through all layer assignments in the small set of test files, and at least there this relationship between layer and individual item was never encountered.

One reason why this 2nd step would be slow is that a LOT OF things in IFC are representation items. Looking at a IfcShellBasedSurfaceModel, all referenced instances: IfcClosedShell - IfcFace - IfcFaceBound - IfcPolyLoop - IfcCartesianPoint are all subclasses of IfcRepresentationItem.

Perhaps an alternative approach would be to set a limit (max_level=2) to `prod->entity->file->traverse(prod->Representation())` so that only IfcProductDefinitionShape - IfcShapeRepresentation - and the immediate items are considered. 

I think that would be a minimal performance difference. I can imagine how for this e.g. MEP file, there are probably huge graphs of items hidden under these representations. But the top one should be a single shell most of the time.

I mean it just feels odd to eliminate code that the schema so explicitly requires: 

~~~
TYPE IfcLayeredItem = SELECT
  ( IfcRepresentationItem
  , IfcRepresentation);
END_TYPE; 
~~~