@Stinkfist0 2019-02-06
Ah, right, for some reason I thought this was already merged! Does the minimal cache invalidation look okay to to you? If yes, I won't be continuing further with this.